comeback to int values because it's better

+++ libs/matrix_lib.h
@@ -7,14 +7,14 @@
 #include "other_lib.h"
 
 //prototypes
-float** readMatrix(char* filename, int rows, int columns);
-void displayMatrix(float** matrix, int rows, int columns);
-float* getColumn(float** matrix, int rows, int column);
-void killColumn(float** matrix, int rows, int columns);
-void swapRows(float** matrix, int a, int b);
+int** readMatrix(char* filename, int rows, int columns);
+void displayMatrix(int** matrix, int rows, int columns);
+int* getColumn(int** matrix, int rows, int column);
+void killColumn(int** matrix, int rows, int columns);
+void swapRows(int** matrix, int a, int b);
 
 //impl
-inline float** readMatrix(char* filename, int rows, int columns){
+inline int** readMatrix(char* filename, int rows, int columns){
     FILE* fp = fopen(filename, "r");
     if (fp == NULL) {
         printf("cannot open %s", filename);
@@ -25,10 +25,10 @@ inline float** readMatrix(char* filename, int rows, int columns){
     for (int i = 0; i < rows; i++)
         *(csv_matrix + i) = readLine(fp, columns);
 
-    float** matrix = malloc(rows * sizeof(float *));
+    int** matrix = malloc(rows * sizeof(int *));
 
     for (int i = 0; i < rows; i++) {
-        matrix[i] = malloc(columns * sizeof(float));
+        matrix[i] = malloc(columns * sizeof(int));
         char** splitted = split(*(csv_matrix + i), ",");
         for (int j = 0; j < columns; j++) {
                matrix[i][j] = atof(*(splitted + j));
@@ -39,7 +39,7 @@ inline float** readMatrix(char* filename, int rows, int columns){
     return matrix;
 }
 
-inline void displayMatrix(float** matrix, int rows, int columns) {
+inline void displayMatrix(int** matrix, int rows, int columns) {
     putchar('\t');
     for (int i = 0; i < columns; i++)
         printf("%d    ", i+1);
@@ -47,24 +47,24 @@ inline void displayMatrix(float** matrix, int rows, int columns) {
     for (int i = 0; i < rows; i++) {
         printf("%d\t", i+1);
         for (int j = 0; j < columns; j++)
-            printf("%.2f    ", matrix[i][j]);
+            printf("%d    ", matrix[i][j]);
         printf("\n\n");
     }
 }
 
-inline float* getColumn(float** matrix, int rows, int column) {
-    float* cols = malloc(rows * sizeof(float));
+inline int* getColumn(int** matrix, int rows, int column) {
+    int* cols = malloc(rows * sizeof(int));
     for (int i = 0; i < rows; i++)
         *(cols + i) = matrix[i][column];
 
     return cols;
 }
 
-inline void killColumn(float** matrix, int rows, int columns) {
-    float pivot = matrix[0][0];
+inline void killColumn(int** matrix, int rows, int columns) {
+    int pivot = matrix[0][0];
     for (int i = 1; i < rows; i++) {
         if (matrix[i][0] != 0) {
-            float toKill = matrix[i][0];
+            int toKill = matrix[i][0];
             for (int j = 0; j < columns; j++) { // iterate on the column of the pivot
                 matrix[i][j] *= pivot;
                 matrix[i][j] -= matrix[0][j] * toKill;
@@ -73,8 +73,8 @@ inline void killColumn(float** matrix, int rows, int columns) {
     }
 }
 
-inline void swapRows(float** matrix, int a, int b){
-    float* swap = *(matrix + a);
+inline void swapRows(int** matrix, int a, int b){
+    int* swap = *(matrix + a);
     *(matrix + a) = *(matrix + b);
     *(matrix + b) = swap;
 }
